Fix OAB linking from HTTP domain to HTTPS
Adding a target="_blank" will open the OAB link in a new tab and should solve the issue of switching from an HTTP domain to a HTTPS domain. Using https://neudorf.github.io/cindieneudorfcounselling/ and then clicking the OAB button works normally on latest versions of Chrome, Safari, and Firefox.